features2d: KAZE: do not use the keypoint.class_id field; use .octave instead
KAZE keypoint detector / feature extractor uses keypoint.class_id
to store "level" parameter (the index in array of evolutions),
and keypoint.octave to store "octave".

However, "octave" is used in a single place, and even there, it
can be obtained as evolution_[level]->octave, so storing it is
redundant.

Therefore, we now store level in the keypoint.octave, and leave
the keypoint.class_id unused, as is suggested by OpenCV docs
for KeyPoint class (i.e., the field is left for user to use)

Note: even if KAZE keypoints were to be combined with other
descriptor extractor, each has its own concept of octaves and
scales, so bringing the octave value out does not really gain
us anything...